Fix BTC/DASH UI issues and rationalize timeouts
- Fix hashrate overflow — ulong cast clamped Bitcoin's ~940 EH/s to ulong.MaxValue (~18.45 EH/s)
- Fix negative amounts and show "Pending" for unconfirmed transactions
- Implement RescanBlockchain for BTC/DASH with confirmation dialog
- Replace magic timeout numbers with named constants in BTC/XMR wallet services
- TransactionDetailsView, AddressBookEntryView: hide XMR-only fields for BTC coins

+++ NervaOneWalletMiner/Rpc/Daemon/Coins/BtcBased/DaemonServiceBaseBTC.cs
@@ -224,7 +224,7 @@ namespace NervaOneWalletMiner.Rpc.Daemon
                                     ResGetBlockchainInfo chainInfo = JsonConvert.DeserializeObject<ResGetBlockchainInfo>(resultToken.ToString())!;
                                     responseObj.Height = chainInfo.blocks;
                                     responseObj.TargetHeight = chainInfo.headers;
-                                    responseObj.NetworkHashRate = (ulong)(chainInfo.difficulty * (MaxTargetConstant / GlobalData.CoinSettings[GlobalData.AppSettings.ActiveCoin].BlockSeconds));
+                                    responseObj.NetworkHashRate = chainInfo.difficulty * (MaxTargetConstant / GlobalData.CoinSettings[GlobalData.AppSettings.ActiveCoin].BlockSeconds);
                                 }
                             }
                         }
